Revert "Move image resource out of experimental (#429)"
This reverts commit bde06030944e588af771d127c8ac7c74cd4c75f7.

+++ internal/framework/subproviders/morpheus/resources/image/read.go
@@ -1,5 +1,7 @@
 // (C) Copyright 2025 Hewlett Packard Enterprise Development LP
 
+//go:build experimental
+
 package image
 
 import (
@@ -85,7 +87,7 @@ func getImageAsState(
 
 	// min_disk
 	if image.MinDisk.Get() != nil {
-		convertToGb := *image.MinDisk.Get() / 1024 / 1024 / 1024
+		convertToGb := int64(*image.MinDisk.Get() / 1024 / 1024 / 1024)
 		state.MinDisk = convert.Int64ToType(&convertToGb)
 	} else {
 		state.MinDisk = basetypes.NewInt64Null()
@@ -93,7 +95,7 @@ func getImageAsState(
 
 	// min_ram
 	if image.MinRam.Get() != nil {
-		convertToGb := *image.MinRam.Get() / 1024 / 1024 / 1024
+		convertToGb := int64(*image.MinRam.Get() / 1024 / 1024 / 1024)
 		state.MinRam = convert.Int64ToType(&convertToGb)
 	} else {
 		state.MinRam = basetypes.NewInt64Null()
